Moved file* endpoints to Blueprint routes
