+++ src/app.py
@@ -13,7 +13,6 @@ from hubmap_sdk import EntitySdk
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
 import argparse
 from pathlib import Path
-from shutil import rmtree # Used by file removal
 from flask import Flask, g, jsonify, abort, request, json, Response
 from flask_cors import CORS
 
@@ -38,29 +37,36 @@ import app_manager
 from dataset import Dataset
 from datacite_doi_helper_object import DataCiteDoiHelper
 
-from app_utils.request_validation import require_json, bad_request_error
+from app_utils.request_validation import require_json
+from app_utils.error import unauthorized_error, not_found_error, internal_server_error, bad_request_error
 from app_utils.misc import __get_dict_prop
 from app_utils.entity import __get_entity
 
 from routes.auth import auth_blueprint
 from routes.datasets import datasets_blueprint
+from routes.file import file_blueprint
 
 # Set logging format and level (default is warning)
 # All the API logging is forwarded to the uWSGI server and gets written into the log file `uwsgi-ingest-api.log`
 # Log rotation is handled via logrotate on the host system with a configuration file
 # Do NOT handle log file and rotation via the Python logging to avoid issues with multi-worker processes
-logging.basicConfig(format='[%(asctime)s] %(levelname)s in %(module)s: %(message)s', level=logging.DEBUG, datefmt='%Y-%m-%d %H:%M:%S')
+logging.basicConfig(format='[%(asctime)s] %(levelname)s in %(module)s: %(message)s',
+                    level=logging.DEBUG,
+                    datefmt='%Y-%m-%d %H:%M:%S')
 logger = logging.getLogger(__name__)
 
 # Specify the absolute path of the instance folder and use the config file relative to the instance path
-app = Flask(__name__, instance_path=os.path.join(os.path.abspath(os.path.dirname(__file__)), 'instance'), instance_relative_config=True)
+app = Flask(__name__,
+            instance_path=os.path.join(os.path.abspath(os.path.dirname(__file__)), 'instance'),
+            instance_relative_config=True)
 app.config.from_pyfile('app.cfg')
 
 app.register_blueprint(auth_blueprint)
 app.register_blueprint(datasets_blueprint)
+app.register_blueprint(file_blueprint)
 
 # Suppress InsecureRequestWarning warning when requesting status on https with ssl cert verify disabled
-requests.packages.urllib3.disable_warnings(category = InsecureRequestWarning)
+requests.packages.urllib3.disable_warnings(category=InsecureRequestWarning)
 
 # Enable/disable CORS from configuration based on docker or non-docker deployment
 if app.config['ENABLE_CORS']:
@@ -256,132 +262,6 @@ def status():
 ####################################################################################################
 ## Ingest API Endpoints
 ####################################################################################################
-
-"""
-File upload handling for Donor and Sample
-
-Returns
--------
-json
-    A JSON containing the temp file id
-"""
-@app.route('/file-upload', methods=['POST'])
-def upload_file():
-    # Check if the post request has the file part
-    if 'file' not in request.files:
-        bad_request_error('No file part')
-
-    file = request.files['file']
-
-    if file.filename == '':
-        bad_request_error('No selected file')
-
-    try:
-        temp_id = file_upload_helper_instance.save_temp_file(file)
-        rspn_data = {
-            "temp_file_id": temp_id
-        }
-
-        return jsonify(rspn_data), 201
-    except Exception as e:
-        # Log the full stack trace, prepend a line with our message
-        msg = "Failed to upload files"
-        logger.exception(msg)
-        internal_server_error(msg)
-
-"""
-File commit triggered by entity-api trigger method for Donor/Sample/Dataset
-
-Donor: image files
-Sample: image files and metadata files
-Dataset: only the one thumbnail file
-
-This call also creates the symbolic from the file uuid dir under uploads
-to the assets dir so the uploaded files can be exposed via gateway's file assets service
-
-Returns
--------
-json
-    A JSON containing the file uuid info
-"""
-@app.route('/file-commit', methods=['POST'])
-def commit_file():
-    # Always expect a json body
-    require_json(request)
-
-    # Parse incoming json string into json data(python dict object)
-    json_data_dict = request.get_json()
-
-    temp_file_id = json_data_dict['temp_file_id']
-    entity_uuid = json_data_dict['entity_uuid']
-    user_token = json_data_dict['user_token']
-
-    file_uuid_info = file_upload_helper_instance.commit_file(temp_file_id, entity_uuid, user_token)
-    filename = file_uuid_info['filename']
-    file_uuid = file_uuid_info['file_uuid']
-
-    # Link the uploaded file uuid dir to assets
-    # /hive/hubmap/hm_uploads/<entity_uuid>/<file_uuid>/<filename> (for PROD)
-    source_file_path = os.path.join(str(app.config['FILE_UPLOAD_DIR']), entity_uuid, file_uuid, filename)
-    # /hive/hubmap/assets/<file_uuid>/<filename> (for PROD)
-    target_file_dir = os.path.join(str(app.config['HUBMAP_WEBSERVICE_FILEPATH']), file_uuid)
-    target_file_path = os.path.join(target_file_dir, filename)
-
-    # Create the file_uuid directory under assets dir
-    # and a symbolic link to the uploaded file
-    try:
-        Path(target_file_dir).mkdir(parents=True, exist_ok=True)
-        os.symlink(source_file_path, target_file_path)
-    except Exception as e:
-        logger.exception(f"Failed to create the symbolic link from {source_file_path} to {target_file_path}")
-
-    # Send back the updated file_uuid_info
-    return jsonify(file_uuid_info)
-
-"""
-File removal triggered by entity-api trigger method for Donor/Sample/Dataset
-during entity update
-
-Donor: image files
-Sample: image files and metadata files
-Dataset: only the one thumbnail file
-
-Returns
--------
-json
-    A JSON list containing the updated files info
-    It's an empty list for Dataset since there's only one thumbnail file
-"""
-@app.route('/file-remove', methods=['POST'])
-def remove_file():
-    # Always expect a json body
-    require_json(request)
-
-    # Parse incoming json string into json data(python dict object)
-    json_data_dict = request.get_json()
-
-    entity_uuid = json_data_dict['entity_uuid']
-    file_uuids = json_data_dict['file_uuids']
-    files_info_list = json_data_dict['files_info_list']
-
-    # `upload_dir` is already normalized with trailing slash
-    entity_upload_dir = file_upload_helper_instance.upload_dir + entity_uuid + os.sep
-
-    # Remove the physical files from the file system
-    for file_uuid in file_uuids:
-        # Get back the updated files_info_list
-        files_info_list = file_upload_helper_instance.remove_file(entity_upload_dir, file_uuid, files_info_list)
-
-        # Also remove the dir contains the symlink to the uploaded file under assets
-        # /hive/hubmap/assets/<file_uuid> (for PROD)
-        assets_file_dir = os.path.join(str(app.config['HUBMAP_WEBSERVICE_FILEPATH']), file_uuid)
-        # Delete an entire directory tree
-        # path must point to a directory (but not a symbolic link to a directory)
-        rmtree(assets_file_dir)
-
-    # Send back the updated files_info_list
-    return jsonify(files_info_list)
-
 
 """
 Retrieve the path of Datasets or Uploads relative to the Globus endpoint mount point give from a list of entity uuids
@@ -1890,38 +1770,6 @@ def validate_donors(headers, records):
 ####################################################################################################
 ## Internal Functions
 ####################################################################################################
-
-
-"""
-Throws error for 401 Unauthorized with message
-Parameters
-----------
-err_msg : str
-    The custom error message to return to end users
-"""
-def unauthorized_error(err_msg):
-    abort(401, description = err_msg)
-
-"""
-Throws error for 404 Not Found with message
-Parameters
-----------
-err_msg : str
-    The custom error message to return to end users
-"""
-def not_found_error(err_msg):
-    abort(404, description = err_msg)
-
-"""
-Throws error for 500 Internal Server Error with message
-Parameters
-----------
-err_msg : str
-    The custom error message to return to end users
-"""
-def internal_server_error(err_msg):
-    abort(500, description = err_msg)
-
 
 # Determines if a dataset is Primary. If the list returned from the neo4j query is empty, the dataset is not primary
 def dataset_is_primary(dataset_uuid):
