+++ src/app_utils/request_validation/__init__.py
@@ -1,5 +1,5 @@
-from flask import abort
 import logging
+from app_utils.error import bad_request_error
 
 logger: logging.Logger = logging.getLogger(__name__)
 
@@ -14,13 +14,3 @@ def require_json(request):
     if not request.is_json:
         bad_request_error("A json body and appropriate Content-Type header are required")
 
-
-def bad_request_error(err_msg):
-    """
-    Throws error for 400 Bad Reqeust with message
-    Parameters
-    ----------
-    err_msg : str
-        The custom error message to return to end users
-    """
-    abort(400, description=err_msg)
